troubleshoot status/config: extract --report rendering into helpers
Move both --report renderers (status ~340 lines, config ~150 lines)
out of custom.py into dedicated helper modules with a single public
entry point, render_report(payload):

  - _troubleshoot_status_report.py
  - _troubleshoot_config_report.py

custom.py now lazily imports each module only on the --report path.
Tests updated to patch the extracted modules where the names are
looked up. No behavioral change.

Copilot-Session: da04b3ec-d5e4-4ab8-aeea-58dd7010ac84
