+++ src/azure-cli/azure/cli/command_modules/appservice/custom.py
@@ -6827,164 +6827,10 @@ def troubleshoot_config(cmd, resource_group_name, name, slot=None, report=False)
         'runtimeError': runtime_error,
     }
     if report:
-        _render_troubleshoot_config(payload)
+        from azure.cli.command_modules.appservice import _troubleshoot_config_report
+        _troubleshoot_config_report.render_report(payload)
         return None
     return payload
-
-
-def _render_troubleshoot_config(payload):
-    """Print the human-readable report: BUILT-IN CHECKS + SITE RUNTIME ERROR
-    RECOMMENDATION. Invoked when ``--report`` is passed."""
-    from azure.cli.core.style import Style, print_styled_text
-
-    def _out(*objs):
-        print_styled_text(*objs, file=sys.stdout)
-
-    def _row(*objs):
-        _out(list(objs))
-
-    def _labeled(label, value):
-        """Emit '<label><value>' with hanging indent so wrapped continuation
-        lines align under the value column."""
-        import shutil
-        import textwrap
-        text = '' if value is None else str(value)
-        term_w = shutil.get_terminal_size(fallback=(120, 40)).columns
-        indent = ' ' * len(label)
-        body_w = max(20, term_w - len(label))
-        lines = textwrap.wrap(text, width=body_w) or [text]
-        _row((Style.PRIMARY, label), (Style.PRIMARY, lines[0]))
-        for cont in lines[1:]:
-            _row((Style.PRIMARY, indent), (Style.PRIMARY, cont))
-
-    config_check = payload.get('configCheck') or {}
-    config_check_failed = payload.get('configCheck') is None
-    config_check_status = payload.get('configCheckStatus')
-    settings = config_check.get('Settings') or config_check.get('settings') or []
-    if not isinstance(settings, list):
-        settings = []
-    settings = [s for s in settings if isinstance(s, dict)]
-    runtime_error = payload.get('runtimeError')
-
-    def _details_level(setting):
-        # KuduLite tags each check with DetailsLevel: 'info' | 'warning' | 'error'.
-        raw = setting.get('DetailsLevel')
-        if raw is None:
-            raw = setting.get('detailsLevel')
-        if isinstance(raw, str):
-            level = raw.strip().lower()
-            if level in ('info', 'warning', 'error'):
-                return level
-        return 'info'
-
-    def _style_for_level(level):
-        if level == 'error':
-            return Style.ERROR
-        if level == 'warning':
-            return Style.WARNING
-        return Style.SUCCESS
-
-    def _is_issue(level):
-        return level in ('warning', 'error')
-
-    any_issue = any(_is_issue(_details_level(s)) for s in settings)
-
-    # Show the runtime error section only when the ARM lastErrorTimestamp is
-    # within the freshness window. The payload already carries the pre-computed
-    # 'isRecent' signal so structured-payload consumers can apply the same gate
-    # (see _RUNTIME_ERROR_FRESHNESS_MINUTES / _runtime_error_is_recent).
-    show_runtime = bool(runtime_error and runtime_error.get('isRecent'))
-
-    # ---- Section 1: Built-in checks ----
-    _out()
-    _row((Style.HIGHLIGHT, '═══ BUILT-IN CHECKS ' + '═' * 55))
-    _out()
-    if config_check_failed:
-        if config_check_status == 404:
-            _row((Style.WARNING, 'Feature is currently unavailable.'))
-        else:
-            _row((Style.WARNING,
-                  'Failed to retrieve built-in configuration checks. Please try again. '
-                  'If the issue persists, restart the application (\'az webapp restart\') and confirm the SCM (Kudu) '
-                  'is running and reachable.'))
-    elif not settings:
-        _row((Style.WARNING, 'No built-in configuration checks reported.'))
-    else:
-        import shutil
-        import textwrap
-        term_w = shutil.get_terminal_size(fallback=(120, 40)).columns
-        setting_w = max(20, min(40, max(len(str(s.get('Setting') or '')) for s in settings) + 2))
-        value_w = max(15, min(30, max(len(str(s.get('Value') or '')) for s in settings) + 2))
-
-        header = '{sname:<{sw}}{vname:<{vw}}{dname}'.format(
-            sname='Setting', sw=setting_w, vname='Value', vw=value_w, dname='Details')
-        _row((Style.HIGHLIGHT, header))
-        _row((Style.SECONDARY, '{s}{v}{d}'.format(
-            s=('─' * (setting_w - 2)).ljust(setting_w),
-            v=('─' * (value_w - 2)).ljust(value_w),
-            d='─' * 40)))
-        for setting in settings:
-            name_v = str(setting.get('Setting') or '')
-            value_v = str(setting.get('Value') if setting.get('Value') is not None else '')
-            details_v = str(setting.get('Details') or '')
-            details_style = _style_for_level(_details_level(setting))
-            prefix = '{s:<{sw}}{v:<{vw}}'.format(
-                s=name_v, sw=setting_w, v=value_v, vw=value_w)
-            details_w = max(20, term_w - len(prefix))
-            detail_lines = textwrap.wrap(details_v, width=details_w) or [details_v]
-            _row((Style.PRIMARY, prefix), (details_style, detail_lines[0]))
-            cont_indent = ' ' * len(prefix)
-            for cont in detail_lines[1:]:
-                _row((Style.PRIMARY, cont_indent), (details_style, cont))
-
-    # ---- Section 2: Site runtime error recommendation ----
-    # Rendered only when the ARM lastErrorTimestamp is within the last 15
-    # minutes. Applied consistently regardless of whether the built-in
-    # checks succeeded, failed, or reported no issues.
-    if show_runtime:
-        _out()
-        _out()
-        _row((Style.HIGHLIGHT, '═══ SITE RUNTIME ERROR RECOMMENDATION ' + '═' * 37))
-        _out()
-        if runtime_error is None:
-            _row((Style.PRIMARY, 'No runtime error reported.'))
-        else:
-            state = str(runtime_error.get('state') or '')
-            last_error = str(runtime_error.get('lastError') or '')
-            details = str(runtime_error.get('lastErrorDetails') or runtime_error.get('details') or '')
-            timestamp_raw = runtime_error.get('lastErrorTimestamp')
-            timestamp = _format_dt(timestamp_raw) or str(timestamp_raw or '')
-            if timestamp:
-                age = _relative_age(timestamp_raw if isinstance(timestamp_raw, str) else None)
-                if age:
-                    timestamp = '{} ({})'.format(timestamp, age)
-
-            if state:
-                _labeled('State                  ', state)
-            if last_error:
-                _labeled('Last Error             ', last_error)
-            if details:
-                _labeled('Last Error Details     ', details)
-            if timestamp:
-                _labeled('Last Error Timestamp   ', timestamp)
-
-        _out()
-
-    # ---- Section 3: Suggested next steps ----
-    # When any built-in check flagged a warning/error, point the user at the
-    # az commands they'll actually need to remediate: update the app setting
-    # and check application logs. Rendered in the same "Hint:" style used by
-    # the status command's failure footer.
-    if any_issue:
-        rg = payload.get('resourceGroup') or '<resource-group>'
-        site_name = payload.get('name') or '<site-name>'
-        _out()
-        _out((Style.WARNING, '▶ Hint:'))
-        _out('  Update flagged app setting:  az webapp config appsettings set -n {} -g {} '
-             '--settings KEY=VALUE'.format(site_name, rg))
-        _out('  Update flagged config:       az webapp config set -n {} -g {} '
-             '--settings KEY=VALUE'.format(site_name, rg))
-        _out('  Check application logs:      az webapp log tail -n {} -g {}'.format(site_name, rg))
 
 
 # ---------------------------------------------------------------------------
@@ -7210,363 +7056,10 @@ def troubleshoot_status(cmd, resource_group, name, slot=None, instance=None, rep
     if orphan_startups:
         payload['orphanStartups'] = orphan_startups
     if report:
-        _render_troubleshoot_status(payload)
+        from azure.cli.command_modules.appservice import _troubleshoot_status_report
+        _troubleshoot_status_report.render_report(payload)
         return None
     return payload
-
-
-def _render_troubleshoot_status(payload):
-    """Print the human-readable report (Site Runtime Status + per-instance Startup summary).
-    Invoked by 'az webapp troubleshoot status' when --report is passed."""
-    from azure.cli.core.style import Style, print_styled_text
-
-    instances = payload.get('instances') or []
-    app_name = payload.get('name') or '<webapp>'
-    resource_group = payload.get('resourceGroup')
-
-    def _out(*objs):
-        print_styled_text(*objs, file=sys.stdout)
-
-    def _row(*objs):
-        _out(list(objs))
-
-    if not instances:
-        _out("No runtime status reported for '{}'.".format(app_name))
-        return
-
-    _out('')
-    _out((Style.HIGHLIGHT, "Application status for {}.".format(app_name)))
-    _out('')
-
-    # Overview table (skip when only one instance is present, e.g. --instance filter).
-    if len(instances) > 1:
-        col_widths = (14, 20, 12, 24)
-        header = "{:<{w0}}{:<{w1}}{:<{w2}}{}".format(
-            'INSTANCE', 'MACHINE', 'STATE', 'UPDATED',
-            w0=col_widths[0], w1=col_widths[1], w2=col_widths[2])
-        _out((Style.PRIMARY, header))
-        _out((Style.PRIMARY, '-' * sum(col_widths)))
-        for inst in instances:
-            startup = inst.get('startup') or {}
-            updated = _format_dt(_most_recent_startup(startup)) or '-'
-            state = inst.get('state') or '-'
-            # Pad plain text first, then wrap the STATE segment in its style — the
-            # framework's color escapes don't perturb the visible column width.
-            _out([
-                (Style.PRIMARY, '{:<{w}}'.format(_short_id(inst.get('instanceId')), w=col_widths[0])),
-                (Style.PRIMARY, '{:<{w}}'.format(inst.get('machineName') or '-', w=col_widths[1])),
-                (_state_style(state), '{:<{w}}'.format(state, w=col_widths[2])),
-                (Style.PRIMARY, updated),
-            ])
-        _out('')
-        _out('')
-
-    # Per-instance Site Runtime Status + Startup summary.
-    for inst in instances:
-        machine = inst.get('machineName')
-        label = machine if machine else _short_id(inst.get('instanceId'))
-        scm_id = inst.get('startupInstanceId')
-        # When ARM's machineName and SCM's InstanceId disagree (common on
-        # Linux App Service — ARM tracks the worker slot, KuduLite tracks the
-        # container) surface both so users can correlate with SCM logs.
-        _out("-" * 76)
-        if scm_id and scm_id != machine:
-            header = 'Instance {} Full Status Report (SCM: {}) '.format(label, scm_id)
-        else:
-            header = 'Instance {} Full Status Report '.format(label)
-        _out(header)
-        _out("-" * 76)
-        #_row((Style.PRIMARY, '═══ ' + header + '══' * 30))
-        _out((Style.HIGHLIGHT, 'Last runtime status'))
-        _print_runtime_block(inst, _out)
-        _out('')
-        _out((Style.HIGHLIGHT, 'Startup summary (last 24h)'))
-        if not machine and not inst.get('startup'):
-            # Without machineName we couldn't query KuduLite for this instance, so
-            # distinguish the "couldn't ask" case from "asked, nothing recorded".
-            _out('  Startup summary unavailable: machine name could not be determined for this instance.')
-            _out('')
-        else:
-            _print_startup_block(inst.get('startup'), _out)
-        _out()
-        _out()
-
-
-    # Orphan startups — SCM entries with no matching ARM instance. These
-    # commonly show up when a container has been recycled: SCM still has the
-    # last container's logs, but ARM has already replaced the worker-slot ID.
-    orphan_startups = payload.get('orphanStartups') or []
-    for orphan in orphan_startups:
-        scm_id = orphan.get('InstanceId') or '<unknown>'
-        _out((Style.HIGHLIGHT, 'Instance {} Startup Summary (no matching ARM instance)'.format(scm_id)))
-        _row((Style.HIGHLIGHT, '─' * 76))
-        _out((Style.HIGHLIGHT, 'Startup summary (last 24h)'))
-        _print_startup_block(orphan.get('Startup'), _out)
-        _out()
-
-    # Hint footer — surfaced only when at least one instance has a real
-    # failure in the report's window (Failed > 0 + lastError set).
-    has_error = any(
-        (inst.get('lastError') and _failed_count(inst.get('startup')) > 0)
-        for inst in instances
-    )
-    if has_error:
-        rg = resource_group or '<resource-group>'
-        _out((Style.WARNING, '▶ Hint:'))
-        _out('  Check application logs:  az webapp log tail -n {} -g {}'.format(app_name, rg))
-        _out('  Check startup logs:      az webapp log startup show -n {} -g {}'.format(app_name, rg))
-
-
-def _state_style(state):
-    """Map a runtime state string to an azure-cli Style for print_styled_text."""
-    from azure.cli.core.style import Style
-    if not state:
-        return Style.PRIMARY
-    s = state.lower()
-    if s == 'started':
-        return Style.SUCCESS
-    if s in ('stopped', 'failed', 'crashed', 'unhealthy'):
-        return Style.ERROR
-    if s in ('starting', 'pullingimage', 'pulling', 'pending'):
-        return Style.WARNING
-    return Style.PRIMARY
-
-
-def _outcome_style(outcome):
-    from azure.cli.core.style import Style
-    if not outcome:
-        return Style.PRIMARY
-    o = outcome.upper()
-    if o == 'STARTED':
-        return Style.SUCCESS
-    if o in ('FAILED', 'CRASHED'):
-        return Style.ERROR
-    return Style.PRIMARY
-
-
-def _count_style(count, kind):
-    """Style for a numeric count. kind='failed' -> ERROR when > 0; 'successful' -> SUCCESS when > 0.
-    Accepts either an int/str integer (e.g. 3, "3") or a KuduLite capped-count
-    string like "50+" (parsed as the leading integer for the > 0 test)."""
-    from azure.cli.core.style import Style
-    text = str(count)
-    try:
-        n = int(text)
-    except (TypeError, ValueError):
-        # Handle capped forms like "50+" — parse the leading digits.
-        import re as _re
-        m = _re.match(r'\d+', text)
-        n = int(m.group(0)) if m else None
-    if n is None:
-        return Style.PRIMARY, text
-    if kind == 'failed' and n > 0:
-        return Style.ERROR, text
-    if kind == 'successful' and n > 0:
-        return Style.SUCCESS, text
-    return Style.PRIMARY, text
-
-
-def _short_id(instance_id):
-    """Truncate a long hex ARM instanceId for table display."""
-    if not instance_id:
-        return '-'
-    if len(instance_id) > 12:
-        return instance_id[:10]
-    return instance_id
-
-
-def _format_dt(value):
-    if not value:
-        return None
-    # Pass through ISO strings; trim sub-second/timezone noise for the table view.
-    if isinstance(value, str):
-        v = value.replace('T', ' ')
-        is_utc = v.endswith('Z')
-        if '.' in v:
-            v = v.split('.', 1)[0]
-        if is_utc:
-            if v.endswith('Z'):
-                v = v[:-1]
-            v = v + ' UTC'
-        elif '+' in v:
-            v = v.split('+', 1)[0]
-        return v
-    return str(value)
-
-
-def _relative_age(iso_value):
-    """Return a short 'Nh Mm ago' / 'Nm ago' / 'just now' / 'in the future' string
-    for an ISO-8601 UTC timestamp, or None if the input is unparseable/missing."""
-    if not iso_value or not isinstance(iso_value, str):
-        return None
-    from datetime import datetime, timezone
-    v = iso_value
-    if '.' in v:
-        # datetime.fromisoformat pre-3.11 chokes on fractional seconds with 'Z' — strip both.
-        head, _, tail = v.partition('.')
-        tz = ''
-        for suffix in ('Z', '+', '-'):
-            if suffix in tail:
-                idx = tail.find(suffix)
-                tz = tail[idx:]
-                break
-        v = head + tz
-    v = v.replace('Z', '+00:00')
-    try:
-        dt = datetime.fromisoformat(v)
-    except ValueError:
-        return None
-    if dt.tzinfo is None:
-        dt = dt.replace(tzinfo=timezone.utc)
-    delta = datetime.now(timezone.utc) - dt
-    total_seconds = int(delta.total_seconds())
-    if total_seconds < 0:
-        return 'in the future'
-    if total_seconds < 60:
-        return 'just now'
-    minutes = total_seconds // 60
-    if minutes < 60:
-        return '{}m ago'.format(minutes)
-    hours = minutes // 60
-    rem_min = minutes % 60
-    if hours < 24:
-        return '{}h {}m ago'.format(hours, rem_min) if rem_min else '{}h ago'.format(hours)
-    days = hours // 24
-    rem_hr = hours % 24
-    return '{}d {}h ago'.format(days, rem_hr) if rem_hr else '{}d ago'.format(days)
-
-
-def _failed_count(startup):
-    """Parse Startup.Failed into an int (0 when missing/invalid). Accepts int,
-    numeric string, or KuduLite's capped '50+' form (leading digits win)."""
-    if not startup:
-        return 0
-    raw = startup.get('Failed')
-    if raw is None:
-        return 0
-    text = str(raw)
-    try:
-        return int(text)
-    except ValueError:
-        import re as _re
-        m = _re.match(r'\d+', text)
-        return int(m.group(0)) if m else 0
-
-
-def _emit_labeled(emit, label, value, value_style=None):
-    """Emit '<label><value>' with hanging indent: if the value wraps at the
-    terminal width, continuation lines are aligned under the value column.
-
-    ``label`` is a pre-padded string (all rows in the same block should use the
-    same width so their values line up). ``emit`` is the same callable the
-    rest of the renderer uses (accepts either a string or a list of styled
-    tuples).
-    """
-    import shutil
-    import textwrap
-    from azure.cli.core.style import Style
-
-    text = value if value is not None else ''
-    if not isinstance(text, str):
-        text = str(text)
-
-    term_w = shutil.get_terminal_size(fallback=(120, 40)).columns
-    indent = ' ' * len(label)
-    body_w = max(20, term_w - len(label))
-    lines = textwrap.wrap(text, width=body_w) or [text]
-
-    if value_style is not None:
-        emit([(Style.PRIMARY, label), (value_style, lines[0])])
-        for cont in lines[1:]:
-            emit([(Style.PRIMARY, indent), (value_style, cont)])
-    else:
-        emit('{}{}'.format(label, lines[0]))
-        for cont in lines[1:]:
-            emit('{}{}'.format(indent, cont))
-
-
-def _print_runtime_block(inst, emit):
-    """Print one Site Runtime Status block from an ARM /siteStatus item."""
-    from azure.cli.core.style import Style
-    if not inst:
-        emit('  (no runtime status reported)')
-        return
-    state = inst.get('state') or '-'
-    details = inst.get('details') or '-'
-    emit([(Style.PRIMARY, '  State                 '), (_state_style(state), state)])
-    _emit_labeled(emit, '  Details               ', details)
-    # LastError may be stale after a recovery, so only surface it when this
-    # worker has actually had failed startup attempts in the report's window.
-    has_visible_error = bool(inst.get('lastError')) and _failed_count(inst.get('startup')) > 0
-    if has_visible_error:
-        last_error = inst.get('lastError') or '-'
-        last_error_details = inst.get('lastErrorDetails') or '-'
-        # Treat .NET DateTime.MinValue (0001-01-01...) as "no error ever" and hide it.
-        last_error_ts_raw = inst.get('lastErrorTimestamp')
-        if isinstance(last_error_ts_raw, str) and last_error_ts_raw.startswith('0001-01-01'):
-            last_error_ts_raw = None
-        last_error_ts = _format_dt(last_error_ts_raw) or '-'
-        age = _relative_age(last_error_ts_raw) if last_error_ts_raw else None
-        if age:
-            last_error_ts = '{} ({})'.format(last_error_ts, age)
-        _emit_labeled(emit, '  Last Error            ', last_error)
-        _emit_labeled(emit, '  Last Error Details    ', last_error_details)
-        _emit_labeled(emit, '  Last Error Timestamp  ', last_error_ts)
-
-
-def _most_recent_startup(startup):
-    """Return the most recent of MostRecentSuccess / MostRecentFailure (ISO strings),
-    or None if both are missing. Lexicographic max is correct for RFC3339/ISO-8601 UTC."""
-    if not startup:
-        return None
-    candidates = [ts for ts in (startup.get('MostRecentSuccess'),
-                                startup.get('MostRecentFailure')) if ts]
-    return max(candidates) if candidates else None
-
-
-def _startup_fetch_failed(startup):
-    """Return the SummaryFetchStatus string only when it indicates a fetch failure
-    (i.e. not the KuduLite success sentinel). None means success or missing."""
-    if not startup:
-        return None
-    status = startup.get('SummaryFetchStatus')
-    if not status:
-        return None
-    # KuduLite success sentinel starts with 'Successfully'; anything else is a
-    # user-facing failure reason we want to surface.
-    if str(status).startswith('Successfully'):
-        return None
-    return status
-
-
-def _print_startup_block(s, emit):
-    from azure.cli.core.style import Style
-    if not s:
-        emit('  No startup attempts recorded in the last 24 hours')
-        emit('')
-        return
-    # KuduLite sets SummaryFetchStatus to a failure reason when it couldn't read
-    # the log directory for this worker; other fields are meaningless then.
-    fetch_error = _startup_fetch_failed(s)
-    if fetch_error:
-        emit([(Style.PRIMARY, '  Startup summary unavailable: '),
-              (Style.WARNING, str(fetch_error))])
-        emit('')
-        return
-    succeeded = s.get('Succeeded', 0)
-    failed = s.get('Failed', 0)
-    emit([(Style.PRIMARY, '  Succeeded              '), _count_style(succeeded, 'successful')])
-    emit([(Style.PRIMARY, '  Failed                 '), _count_style(failed, 'failed')])
-    most_recent_success = _format_dt(s.get('MostRecentSuccess'))
-    most_recent_failure = _format_dt(s.get('MostRecentFailure'))
-    emit([(Style.PRIMARY, '  Most recent success    '),
-          (_outcome_style('STARTED') if most_recent_success else Style.PRIMARY,
-           most_recent_success or '-')])
-    emit([(Style.PRIMARY, '  Most recent failure    '),
-          (_outcome_style('FAILED') if most_recent_failure else Style.PRIMARY,
-           most_recent_failure or '-')])
-    emit('')
-
 
 
 def config_slot_auto_swap(cmd, resource_group_name, webapp, slot, auto_swap_slot=None, disable=None):
